add: chatroom member to chatroom

+++ src/modules/chat/usecase/delete_scheduled_message.ts
@@ -1,4 +1,5 @@
 import { UniqueID } from "../../../shared/domain/UniqueID";
+import { ChatRoom } from "../domain/ChatRoom";
 import { ScheduledChatMessageRepo } from "../infra/repository/ScheduledMessage"
 
 export class DeleteScheduledMessage {
@@ -8,7 +9,11 @@ export class DeleteScheduledMessage {
     this.repo = repo
   }
 
-  do(id: UniqueID) {
+  do(chatRoom: ChatRoom, id: UniqueID) {
+    const msg = this.repo.get(id)
+    if (!chatRoom.memberExist(msg.userId)) {
+      throw "member not include in chatroom"
+    }
     this.repo.delete(id)
   }
 }
